Moved creation of vtk rendering pipelines to render thread.
The main design flaw of V0.1 was the fact that the rendering
pipeline was created in the main thread, preventing interactive
communication from the render thread. Now the render thread
triggers the pipeline construction if it sees no actors in a
given figure. In this case it calls the Build() method of this
figure. This allows the render thread to influence pipeline
data.

For an example, see the slider in vtkfig::Contour2D.